requirements.txt
requests
python-dotenv
pandas
scikit-learn
flask
gunicorn
boto3
